fluxor implementation

+++ ChatApp.Server/Security/Interfaces/IConnectedUser.cs
@@ -2,5 +2,5 @@ namespace ChatApp.Server.Security.Interfaces;
 
 public interface IConnectedUser
 {
-    public int? Id { get; }
+    public long? Id { get; }
 }
